# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Xet Server API
  version: 0.8.0
  description: |
    HTTP APIs exposed by `xetd`: a wire-compatible Xet CAS server
    (`internal/casserver`, mounted at `/v1` and `/v2`), an optional Hub API
    shim (`internal/hubserver`, mounted on a separate port via `-hub-addr`)
    that lets the real `hf` CLI talk to this server end-to-end, and the Xet
    Data API (`internal/api`, mounted at `/v1` alongside the CAS protocol,
    on a disjoint set of literal paths) - a simple, non-wire-compatible
    chunk/dedup API for quick manual testing via the `xet` CLI.

    ## Authentication

    Every endpoint except telemetry/storage-stats (CAS) and `/v1/stats`
    (Xet Data API) is gated by a pluggable `auth.Authenticator`,
    defaulting to no enforcement at all (this server's pre-v0.8.0
    behavior). Enabled via `xetd -auth-token <secret>` (or
    `$XETD_AUTH_TOKEN`/`$HF_TOKEN`), which installs a single shared bearer
    token checked against every request:

    - **401 Unauthorized** - no `Authorization` header, or a token that
      fails validation entirely (wrong token, malformed header).
    - **403 Forbidden** - a valid token that lacks the scope (`read` or
      `write`) the endpoint requires.

    In this UI, use **Authorize** and paste the same secret the server was
    started with; it is sent as `Authorization: Bearer <token>` on every
    "Try it out" call. If the server has no auth configured you can skip
...

    ## Trying endpoints from this page

...

...

...

...
servers:
...
security:
  - bearerAuth: []
components:
  responses:
    Unauthorized:
      description: No credential, or one that fails validation entirely.
      content:
        text/plain:
          schema: { type: string }
    Forbidden:
      description: Valid credential, but lacking the scope this endpoint requires.
      content:
        text/plain:
          schema: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Optional. Only enforced if xetd was started with -auth-token (or
        $XETD_AUTH_TOKEN/$HF_TOKEN) set to something other than "None". A
        server with no auth configured accepts every request regardless of
        whether this header is present.
  schemas:
...
    UploadResult:
      type: object
      properties:
        file_id: { type: string }
        name: { type: string }
        size: { type: integer, format: int64 }
        sha256: { type: string }
        chunks_total: { type: integer }
        chunks_new: { type: integer }
        bytes_stored: { type: integer, format: int64 }
        bytes_on_wire: { type: integer, format: int64 }
        dedup_percent:
          type: number
          format: double
          description: '100 * (1 - bytes_stored/size); 0 if size is 0.'
      required:
        - file_id
        - size
        - sha256
        - chunks_total
        - chunks_new
        - bytes_stored
        - bytes_on_wire
        - dedup_percent
    ChunkRef:
      type: object
      properties:
        hash: { type: string }
        offset: { type: integer, format: int64 }
        length: { type: integer }
      required: [hash, offset, length]
    Manifest:
      type: object
      properties:
        file_id: { type: string }
        size: { type: integer, format: int64 }
        sha256: { type: string }
        chunks:
          type: array
          items: { $ref: '#/components/schemas/ChunkRef' }
      required: [file_id, size, sha256, chunks]
...
paths:
...
  /v1/upload:
    post:
      tags: [Xet Data]
      summary: Chunk, dedup, and store a file
      description: Requires write scope.
      parameters:
        - { name: name, in: query, required: false, schema: { type: string }, description: Echoed back in the response; not used for storage/hashing. }
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema: { type: string, format: binary }
      responses:
        '200':
          description: Upload summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UploadResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500':
          description: Storage or manifest write failure.
  /v1/files/{id}:
    get:
      tags: [Xet Data]
      summary: Download a previously-uploaded file
      description: >-
        Requires read scope. No Range header support - always streams the
        full file.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Full file bytes.
          headers:
            Content-Length: { schema: { type: integer, format: int64 } }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown or invalid file id.
        '500':
          description: A chunk referenced by the manifest is missing.
  /v1/files/{id}/manifest:
    get:
      tags: [Xet Data]
      summary: Fetch a file's chunk manifest
      description: Requires read scope.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Manifest.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Manifest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Unknown or invalid file id.
...